Create common base classes for common functionality
Lift common functionality into a ReaderCommon class providing the
version and datamodel information. TTree based stuff starts to move into
a TTreeReaderBase
