+++ include/podio/ROOTLazyReader.h
@@ -4,6 +4,8 @@
 #include "podio/ROOTLazyFrameData.h"
 #include "podio/podioVersion.h"
 #include "podio/utilities/DatamodelRegistryIOHelpers.h"
+#include "podio/utilities/ReaderCommon.h"
+#include "podio/utilities/RootHelpers.h"
 
 #include "TChain.h"
 
@@ -26,7 +28,7 @@ class GenericParameters;
 /// podio::Frame can be constructed. It can be used to read files written by the
 /// ROOTWriter. Unlike the ROOTReader which reads all collections eagerly, this
 /// reader defers the actual ROOT I/O to the point of collection access.
-class ROOTLazyReader {
+class ROOTLazyReader : public ReaderCommon, root_utils::TTreeReaderCommon {
 
 public:
   ROOTLazyReader() = default;
@@ -93,45 +95,10 @@ public:
   /// @returns The number of entries that are available for the category
   unsigned getEntries(std::string_view name) const;
 
-  /// Get the build version of podio that has been used to write the current
-  /// file
-  ///
-  /// @returns The podio build version
-  podio::version::Version currentFileVersion() const {
-    return m_fileVersion;
-  }
-
-  /// Get the (build) version of a datamodel that has been used to write the
-  /// current file
-  ///
-  /// @param name The name of the datamodel
-  ///
-  /// @returns The (build) version of the datamodel if available or an empty
-  ///          optional
-  std::optional<podio::version::Version> currentFileVersion(std::string_view name) const {
-    return m_datamodelHolder.getDatamodelVersion(name);
-  }
-
   /// Get the names of all the available Frame categories in the current file(s).
   ///
   /// @returns The names of the available categories from the file
   std::vector<std::string_view> getAvailableCategories() const;
-
-  /// Get the datamodel definition for the given name
-  ///
-  /// @param name The name of the datamodel
-  ///
-  /// @returns The high level definition of the datamodel in JSON format
-  const std::string_view getDatamodelDefinition(std::string_view name) const {
-    return m_datamodelHolder.getDatamodelDefinition(name);
-  }
-
-  /// Get all names of the datamodels that are available from this reader
-  ///
-  /// @returns The names of the datamodels
-  std::vector<std::string> getAvailableDatamodels() const {
-    return m_datamodelHolder.getAvailableDatamodels();
-  }
 
 private:
   /// Initialize the passed CategoryState by setting up the necessary branches,
@@ -152,9 +119,6 @@ private:
   std::unique_ptr<TChain> m_metaChain{nullptr};
   std::unordered_map<std::string_view, std::shared_ptr<CategoryState>> m_categoryStates{};
   std::vector<std::string> m_availCategories{};
-
-  podio::version::Version m_fileVersion{0, 0, 0};
-  DatamodelDefinitionHolder m_datamodelHolder{};
 };
 
 } // namespace podio
