+++ src/rootUtils.h
@@ -409,102 +409,6 @@ inline std::shared_ptr<podio::CollectionIDTable> makeCollIdTable(const std::vect
   return std::make_shared<podio::CollectionIDTable>(std::move(ids), std::move(names));
 }
 
-inline std::tuple<std::vector<root_utils::CollectionBranches>, std::vector<detail::NamedCollInfo>>
-createCollectionBranches(TChain* chain, const podio::CollectionIDTable& idTable,
-                         const std::vector<root_utils::CollectionWriteInfo>& collInfo) {
-  size_t collectionIndex{0};
-  std::vector<root_utils::CollectionBranches> collBranches;
-  collBranches.reserve(collInfo.size() + 1);
-  std::vector<detail::NamedCollInfo> storedClasses;
-  storedClasses.reserve(collInfo.size());
-
-  for (const auto& [collID, collType, isSubsetColl, collSchemaVersion, _, __] : collInfo) {
-    // We only write collections that are in the collectionIDTable, so no need
-    // to check here
-    const auto name = idTable.name(collID).value();
-
-    root_utils::CollectionBranches branches{};
-    if (isSubsetColl) {
-      // Only one branch will exist and we can trivially get its name
-      const auto brName = root_utils::subsetBranch(name);
-      branches.refs.push_back(root_utils::getBranch(chain, brName.c_str()));
-      branches.refNames.emplace_back(std::move(brName));
-    } else {
-      // This branch is guaranteed to exist since only collections that are
-      // also written to file are in the info metadata that we work with here
-      branches.data = root_utils::getBranch(chain, name.c_str());
-
-      const auto relVecNames = podio::DatamodelRegistry::instance().getRelationNames(collType);
-      for (const auto& relName : relVecNames.relations) {
-        const auto brName = root_utils::refBranch(name, relName);
-        branches.refs.push_back(root_utils::getBranch(chain, brName.c_str()));
-        branches.refNames.emplace_back(std::move(brName));
-      }
-      for (const auto& vecName : relVecNames.vectorMembers) {
-        const auto brName = root_utils::refBranch(name, vecName);
-        branches.vecs.push_back(root_utils::getBranch(chain, brName.c_str()));
-        branches.vecNames.emplace_back(std::move(brName));
-      }
-    }
-
-    storedClasses.emplace_back(name, std::make_tuple(collType, isSubsetColl, collSchemaVersion, collectionIndex++));
-    collBranches.emplace_back(std::move(branches));
-  }
-
-  return {std::move(collBranches), storedClasses};
-}
-
-inline std::tuple<std::vector<root_utils::CollectionBranches>, std::vector<detail::NamedCollInfo>>
-createCollectionBranchesIndexBased(TChain* chain, const podio::CollectionIDTable& idTable,
-                                   const std::vector<root_utils::CollectionWriteInfo>& collInfo) {
-  size_t collectionIndex{0};
-  std::vector<root_utils::CollectionBranches> collBranches;
-  collBranches.reserve(collInfo.size() + 1);
-  std::vector<detail::NamedCollInfo> storedClasses;
-  storedClasses.reserve(collInfo.size());
-
-  for (const auto& [collID, collType, isSubsetColl, collSchemaVersion, _, __] : collInfo) {
-    // We only write collections that are in the collectionIDTable, so no need
-    // to check here
-    const auto name = idTable.name(collID).value();
-
-    const auto collectionClass = TClass::GetClass(collType.c_str());
-    // Need the collection here to setup all the branches. Have to manage the
-    // temporary collection ourselves
-    const auto collection =
-        std::unique_ptr<podio::CollectionBase>(static_cast<podio::CollectionBase*>(collectionClass->New()));
-    root_utils::CollectionBranches branches{};
-    if (isSubsetColl) {
-      // Only one branch will exist and we can trivially get its name
-      const auto brName = root_utils::refBranch(name, 0);
-      branches.refs.push_back(root_utils::getBranch(chain, brName.c_str()));
-      branches.refNames.emplace_back(std::move(brName));
-    } else {
-      // This branch is guaranteed to exist since only collections that are
-      // also written to file are in the info metadata that we work with here
-      branches.data = root_utils::getBranch(chain, name.c_str());
-
-      const auto buffers = collection->getBuffers();
-      for (size_t i = 0; i < buffers.references->size(); ++i) {
-        const auto brName = root_utils::refBranch(name, i);
-        branches.refs.push_back(root_utils::getBranch(chain, brName.c_str()));
-        branches.refNames.emplace_back(std::move(brName));
-      }
-
-      for (size_t i = 0; i < buffers.vectorMembers->size(); ++i) {
-        const auto brName = root_utils::vecBranch(name, i);
-        branches.vecs.push_back(root_utils::getBranch(chain, brName.c_str()));
-        branches.vecNames.emplace_back(std::move(brName));
-      }
-    }
-
-    storedClasses.emplace_back(name, std::make_tuple(collType, isSubsetColl, collSchemaVersion, collectionIndex++));
-    collBranches.emplace_back(std::move(branches));
-  }
-
-  return {std::move(collBranches), storedClasses};
-}
-
 inline std::vector<std::string> getAvailableCategories(TChain* metaChain) {
   const auto* branches = metaChain->GetListOfBranches();
   std::vector<std::string> brNames;
